LayoutTests/fast/webgpu/regression/repro_173266457-expected.txt

CONSOLE MESSAGE: Index buffer contents after submission:

PASS Out-of-order command buffer submission: index buffer contains OOB values after B submitted before A
